adding comments throughout the code
