+++ ts/components/wallet/CreditCardComponent.tsx
@@ -1,3 +1,16 @@
+/**
+ * This component is used to display a credit card 
+ * with its information. 
+ */
+// TODO: as part of #157422715 (CreditCardComponent story), this component will be substantially
+// modified, in order to:
+// - have a grid-only layout that does not rely on margins and other ungodly things
+// - be used throughout the app whenever needed (as of now, CreditCardComponent can
+//   only be used when displaying the list of credit cards, with the rest being 
+//   cropped pictures)
+// TODO: at some point, this component may need to be restructured
+// to display different payment methods
+
 import * as React from "react";
 
 import { Body, Card, Icon, Text } from "native-base";
