+++ ts/components/wallet/layout/Subtitle.tsx
@@ -1,3 +1,7 @@
+/**
+ * Component representing a subtitle within TopContents.
+ */
+
 import { Text } from "native-base";
 import * as React from "react";
 import { Grid, Row } from "react-native-easy-grid";
@@ -8,6 +12,7 @@ type Props = Readonly<{
   content: TopContent;
 }>;
 
+/* size of the subtitle row */
 export const SUBTITLE_SIZE = 2;
 
 export class Subtitle extends React.Component<Props> {
