+++ ts/components/wallet/layout/SubtitlesLR.tsx
@@ -1,3 +1,9 @@
+/**
+ * Component representing two subtitles within a TopContents:
+ * these two subtitles are left- and right-aligned within the
+ * top part of the layout
+ */
+
 import { Left, Right, Text } from "native-base";
 import * as React from "react";
 import { Col, Grid, Row } from "react-native-easy-grid";
@@ -10,12 +16,19 @@ type Props = Readonly<{
   navigation: NavigationScreenProp<NavigationState>;
 }>;
 
+/* size of the component ("height") */
 export const SUBTITLES_LR_SIZE = 2;
 
 export class SubtitlesLR extends React.Component<Props> {
   public render(): React.ReactNode {
     const { subtitleLeftText, subtitleRightText } = this.props
       .content as TopContentSubtitlesLR;
+    /* the left part of the component should be touchable,
+     * right now the destination is hardcoded since there
+     * is no use for additional destinations. In the future, 
+     * this may no longer be the case and a touchable content
+     * may need to be provided via props
+     */
     return (
       <Row size={SUBTITLES_LR_SIZE}>
         <Grid>
