+++ ts/components/wallet/layout/Title.tsx
@@ -1,15 +1,21 @@
+/**
+ * Component representing the title in a TopContents. 
+ * The title is comprised by a text and an optional
+ * image displayed on the right-hand side of the screen
+ */
 import { H1 } from "native-base";
 import * as React from "react";
 import { Image } from "react-native";
 import { Col, Grid, Row } from "react-native-easy-grid";
 import { WalletStyles } from "../../styles";
-import { ImageType } from "./PayLayout";
+import { ImageType } from "./WalletLayout";
 
 type Props = Readonly<{
   text: string;
   rightImage?: ImageType;
 }>;
 
+/* size of the title component (in "rows") */
 export const TITLE_SIZE = 3;
 
 export class Title extends React.Component<Props> {
@@ -17,6 +23,7 @@ export class Title extends React.Component<Props> {
     const images = {
       [ImageType.BANK_IMAGE]: require("../../../../img/wallet/wallet-icon.png")
     };
+    /* optionally display an image, if specified */
     if (this.props.rightImage !== undefined) {
       return (
         <Col size={1}>
