+++ ts/components/wallet/layout/types.ts
@@ -1,4 +1,11 @@
-/* subtitlesLS and subtitle are mutually exclusive */
+/**
+ * TopContent defines the type of top content
+ * that needs to be displayed by WalletLayout. 
+ * This can be a combination of the following
+ * With* and Without*
+ */
+
+/* subtitlesLR and subtitle are mutually exclusive */
 type WithSubtitle = {
   hasMainSubtitle: true;
   hasSubtitlesLR: false;
@@ -26,6 +33,10 @@ type WithoutTouchable = {
   hasTouchable: false;
 };
 
+/* these are the possible options for TopContent,
+ * based on whether the specific parts are present
+ * or not -- the code should be self-explanatory
+ */
 export type TopContentSubtitle = WithSubtitle & WithoutTouchable;
 export type TopContentSubtitlesLR = WithSubtitlesLR & WithoutTouchable;
 export type TopContentTouchable = WithoutSubtitles & WithTouchable;
@@ -34,6 +45,12 @@ export type TopContentSubtitleTouchable = WithSubtitle & WithTouchable;
 export type TopContentSubtitlesLRTouchable = WithSubtitlesLR & WithTouchable;
 export type TopContentNone = WithoutSubtitles & WithoutTouchable;
 
+/* 
+ * The following functions create objects
+ * of the appropriate types with the
+ * approparite has* tags already set, and
+ * integrating with the required data
+ */
 export function topContentSubtitle(subtitleText: string): TopContentSubtitle {
   return {
     hasMainSubtitle: true,
@@ -103,6 +120,9 @@ export function topContentNone(): TopContentNone {
   };
 }
 
+/* TopContent can be either of the
+ * following possible TopContent-s
+ */
 export type TopContent =
   | TopContentSubtitle
   | TopContentSubtitleTouchable
